requirements.txt
# Core dependencies (install in this order)

# 1. PyTorch — install first from https://pytorch.org/get-started/locally/
#    Example (macOS, CPU/MPS):
#      pip install torch torchaudio

# 2. NeMo ASR toolkit
nemo_toolkit[asr]>=2.2.0

# 3. App dependencies
sounddevice>=0.4.6
numpy>=1.26
pynput>=1.7.6
rumps>=0.4.0
